Add querying balance by telegram username

+++ src/MatatakiBot/Services/IBackendService.cs
@@ -6,7 +6,7 @@ namespace MatatakiBot.Services
     public interface IBackendService
     {
         ValueTask<UserInfo> GetUserAsync(int id);
-        ValueTask<UserInfo> GetUserByTelegramIdAsync(int id);
+        ValueTask<UserInfo> GetUserByTelegramIdAsync(long id);
         ValueTask<TokenInfo> GetTokenAsync(string symbol);
     }
 }
