+++ src/MatatakiBot/Services/Impls/BackendService.cs
@@ -20,7 +20,7 @@ namespace MatatakiBot.Services.Impls
 
             return wrapper.Data;
         }
-        public async ValueTask<UserInfo> GetUserByTelegramIdAsync(int id)
+        public async ValueTask<UserInfo> GetUserByTelegramIdAsync(long id)
         {
             var wrapper = await _httpClient.GetFromJsonAsync<ApiWrapper<UserInfo>>("/mapping/telegramUidToUser/" + id);
